gcp/requirements.txt
tensorflow
google-cloud-storage==1.16.1
pillow==6.0.0